refactor: added base screen
